+++ Code/ESP32/dsp/main/circ-buffer.h
@@ -12,8 +12,8 @@ class Circ_Buffer {
         uint32_t writeSpaceAvail() const;
         uint8_t *getReadPointer();
 
-        bool write( const uint8_t *source, uint32_t dataToWrite );
-        bool read( uint32_t readSize, uint8_t *&sourcePtr1, uint32_t &dataSize1, uint8_t *&sourcePtr2, uint32_t &dataSize2 );
+        bool write( const uint8_t *source, size_t dataToWrite );
+        bool read( uint32_t readSize, uint8_t *&sourcePtr1, size_t &dataSize1, uint8_t *&sourcePtr2, size_t &dataSize2 );
     protected:
         uint8_t mBufferSize;
         uint8_t mDataReady;
